requirements.txt
about-time==4.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alive-progress==3.2.0
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==24.3.0
autograd==1.7.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
captcha==0.7.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
clarabel==0.10.0
click==8.1.8
cloudpickle==3.1.1
cma==3.2.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cvxpy==1.6.4
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.9
ecos==2.0.14
executing==2.2.0
extra-streamlit-components==0.1.71
filelock==3.18.0
fonttools==4.56.0
fredapi==0.5.2
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grapheme==0.6.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.62.3
html5lib==1.1
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
narwhals==1.21.1
nest-asyncio==1.6.0
numpy==1.26.4
osqp==0.6.7.post3
packaging==23.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pillow==10.4.0
platformdirs==4.3.6
plotly==5.19.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==4.25.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymoo==0.6.1.3
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyportfolioopt==1.5.6
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
qdldl==0.1.7.post5
referencing==0.35.1
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scipy==1.15.2
scs==3.2.7.post2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.14.0
snowflake-sqlalchemy==1.7.3
sortedcontainers==2.4.0
soupsieve==2.6
spyder-kernels==3.0.3
SQLAlchemy==2.0.39
stack-data==0.6.3
streamlit==1.43.2
streamlit-authenticator==0.4.2
tabulate==0.9.0
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
XlsxWriter==3.2.2
yarl==1.18.3
yfinance==0.2.54
